backend/requirements.txt
Flask-Reuploaded==1.4.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
numpy==1.26.4
pandas==2.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
setuptools==68.2.2
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.10.0
tzdata==2024.1
Werkzeug==3.0.2
wheel==0.41.2
psycopg2-binary==2.9.9
flask_migrate==4.0.7
